pyproject.toml
[tool.poetry]
name = "aharas-de3-1-5"
version = "0.1.0"
description = ""
authors = ["hvsio <alicjaharaszczuk@gmail.com>"]
readme = "README.md"
packages = [{include = "aharas_de3"}]

[tool.poetry.dependencies]
python = "^3.10"
kaggle = "^1.6.6"
sqlalchemy = "^2.0.28"
pandas = "^2.2.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
kubernetes = "^29.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
